CLOUDP-156370: use process hostnames (#2739)
* CLOUDP-156370 Use process.hostname


* Remove ephemeralPortOffset for single-cluster

* Controller logic fix

* Start kind clusters sequentially

* Updated logic in the agent launcher

* Comments addressed

---------

+++ docker/mongodb-enterprise-init-database/content/agent-launcher-lib.sh
@@ -73,7 +73,10 @@ ensure_certs_symlinks() {
         fi
     else
         script_log "Using Kubernetes CA file"
-        ln -sf "/var/run/secrets/kubernetes.io/serviceaccount/ca.crt" "${pod_secrets_dir}/ca.pem"
+
+        if [[ ! -f "${pod_secrets_dir}/ca.pem" ]]; then
+          ln -sf "/var/run/secrets/kubernetes.io/serviceaccount/ca.crt" "${pod_secrets_dir}/ca.pem"
+        fi
     fi
 }
